requirements.txt
alembic==1.4.2
attrs==21.2.0
billiard==3.6.4.0
bcrypt==3.2.0
celery==4.4.7
cffi==1.14.6
cic-eth~=0.13.0
cic-notify~=0.4.10
cic-translations~=0.0.3
cic-types~=0.2.8
phonenumbers==8.12.12
psycopg2==2.8.6
python-i18n[YAML]==0.3.9
pyxdg==0.27
redis==3.5.3
semver==2.13.0
SQLAlchemy==1.3.20
tinydb==4.2.0
transitions==0.8.4
uWSGI==2.0.19.1
backports.zoneinfo==0.2.1;python_version<"3.9"